File_extractor/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-magic-bin; sys_platform == 'win32'
python-magic; sys_platform != 'win32'
pdfplumber
python-docx
pytesseract
opencv-python-headless
numpy
pillow
pandas
beautifulsoup4
pydantic
